+++ lib/src/BackendSettings/appsystemsettings.dart
@@ -1,0 +1,8 @@
+import 'package:interactivesso_datatypes/interactivesso_datatypes.dart';
+
+class BackendAPPSystemSettings{
+  final APPPermissionInfo appBasePermissionInfo;
+  BackendAPPSystemSettings({
+    required this.appBasePermissionInfo
+  });
+}