+++ lib/src/BackendSettings/usersystemsettings.dart
@@ -1,0 +1,8 @@
+import 'package:interactivesso_datatypes/interactivesso_datatypes.dart';
+
+class BackendUserSystemSettings{
+  final UserPermissionInfo userBasePermission;
+  BackendUserSystemSettings({
+    required this.userBasePermission
+  });
+}